Add files via upload
We updated both the Arduino Sketch and the Max MSP patch so that if the wireless connection fails, or if the battery dies, the brightness messages being sent from Max MSP can be received by the ESP via serial. This was the last thing that needs updating so that the GLOBE can be used entirely wirelessly, or plugged in,  without having to re-upload a new Arduino sketch.