Add files via upload
In the previous versions the colour data was sent directly from the TCS34725 colour sensor to the LEDs. This meant no matter what, as long as some sort of colour is pressed against the sensor, you will see it. However, in Max MSP, it is a bit more sensitive to the exact shade of colour and angle the colour is pressed against the sensor. Therefore, if lets say I am using a green sample, the LEDs will appear green right away, but it might not quite be the green that the Max Patch is programmed to trigger the "green" message. It would take slight adjustments for the performer to do to get it accurate. It is the equivalent of have a violin that no matter what played perfect pitch even if the fingers are technically out of tune, and everything else is following the out of tune fingers. 

Long story short, we changed the Arduino sketch so that the RGB and brightness values are all sent from Max MSP to control the LEDs. This way everything is in sync and the LEDs truly reflect what the performer is both hearing and seeing. This makes it easier for the performer to use this feedback to adjust and correct their own movements in real time. This opens up more possibilities for the LEDs to be audio reactive as well. I included a screenshot example for how this is hooked up in Max MSP. 

In previous versions the RGB values outputted from the TCS34725 colour sensor were also well over 500. In the past I never questioned it because it was copied from an example sketch and as long as I was receiving something I thought it was fine. As I have been working more deeply with colour lately, I realized that can't be right - typically RGB values are only between 0-255. Kevin and I went over the code and we found where the issue was. So now the values sent from the sensor to Max are correct. I had updated my colour detection Max patch to reflect this.